Fix error in RepeatListener
