+++ app/src/main/java/de/jbamberger/irremote/remote/IrRemoteProvider.kt
@@ -1,33 +1,10 @@
 package de.jbamberger.irremote.remote
 
 import android.content.Context
-import android.content.pm.PackageManager
 import android.hardware.ConsumerIrManager
 import de.jbamberger.irremote.R
-import timber.log.Timber
 
 class MissingHardwareFeatureException(message: String?) : Exception(message)
-
-data class IrRemote(
-        private val remoteDef: RemoteDefinition,
-        private val irManager: ConsumerIrManager,
-        private val vibrator: VibratorWrapper) {
-
-    val layout: IrRemoteLayout
-        get() = remoteDef.layout
-    val commandDefs: IrCommandSet
-        get() = remoteDef.commandDefs
-
-    fun sendCommand(command: String) {
-        val code = remoteDef.commandDefs.codeMap[command]
-                ?: throw IllegalArgumentException("Unknown command name $command.")
-        val frequency = remoteDef.commandDefs.frequency
-
-        vibrator.vibrate()
-        irManager.transmit(frequency, code)
-        Timber.d("The required frequency range is not supported.")
-    }
-}
 
 class IrRemoteProvider(context: Context) {
